#1970 Fixed httpSenderTest - moved valid url to HttpSenderVO.validate;
