+++ WebContent/WEB-INF/jsp/publisherEdit/editHttpSender.jsp
@@ -26,9 +26,10 @@
   
   function init() {
       PublisherEditDwr.initSender(initCB);
+      initInputSelect();
   }
   dojo.addOnLoad(init);
-  
+
   function initCB(response) {
       var i;
       var list = response.data.allPoints;
@@ -253,8 +254,7 @@
       showMessage("httpSendTestMessage", "<fmt:message key="publisherEdit.httpSender.sending"/>");
       showMessage("httpSendTestData");
       httpSendTestButtons(true);
-      PublisherEditDwr.httpSenderTest($get("url"), $get("usePost") == "true", staticHeaderList, staticParameterList,
-    		  httpSendTestCB);
+      PublisherEditDwr.httpSenderTest(httpSendTestCB);
   }
   
   function httpSendTestButtons(sending) {
@@ -291,9 +291,42 @@
       httpSendTestButtons(false);
       showMessage("httpSendTestMessage", "<fmt:message key="common.cancelled"/>");
   }
+
+  function initInputSelect() {
+      initHttpSenderTest();
+      var inputs = document.querySelectorAll('#publisherEditor input');
+      var selects = document.querySelectorAll('#publisherEditor select');
+      addChangeEvent(inputs);
+      addChangeEvent(selects);
+  }
+
+  function addChangeEvent(tab) {
+      if(tab && tab.forEach) {
+        tab.forEach(function(currentValue, currentIndex, listObj) {
+            console.log('currentValue:' + currentValue);
+            if(currentValue.type && currentValue.type != 'button')
+              currentValue.addEventListener('change', function() { hideHttpSenderTest(); }, false);
+            }, '');
+      }
+  }
+
+  function initHttpSenderTest() {
+    if(${not empty publisher and publisher.id != -1})
+        showHttpSenderTest();
+    else
+        hideHttpSenderTest();
+  }
+
+  function hideHttpSenderTest() {
+      document.getElementById("httpSenderTest").style.visibility = "hidden";
+  }
+
+  function showHttpSenderTest() {
+      document.getElementById("httpSenderTest").style.visibility = "visible";
+  }
 </script>
 
-<table cellpadding="0" cellspacing="0">
+<table id="publisherEditor" cellpadding="0" cellspacing="0">
   <tr>
     <td valign="top">
       <div class="borderDiv marR marB">
@@ -365,9 +398,9 @@
         </table>
       </div>
     </td>
-    
+
     <td valign="top">
-      <div class="borderDiv marB">
+      <div id="httpSenderTest" class="borderDiv marB">
         <table>
           <tr><td class="smallTitle"><fmt:message key="publisherEdit.httpSender.sendTest"/></td></tr>
           <tr>
